pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "modesolver"
version = "2.1.0"
description = "Vector Finite Difference Modesolver for Anisotropic Dielectric Waveguides"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
    { name = "Thomas E. Murphy" }
]
dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
]

[project.optional-dependencies]
fast = [
    "pypardiso",
    "python-mumps>=0.0.6",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["modesolver*"]
